requirements.txt
flask
fpdf
python-docx
pandas
openpyxl
